# speedup derivation parsing by optimizing `parseString`