<!--

IMPORTANT

Nix is a non-trivial project, so for your contribution to be successful,
it really is important to follow the contributing guidelines:

https://github.com/NixOS/nix/blob/master/CONTRIBUTING.md

Even if you've contributed to open source before, take a moment to read it,
so you understand the process and the expectations.

- what information to include in commit messages
- proper attribution
- volunteering contributions effectively
- how to get help and our review process.

-->

## Motivation

<!-- Briefly explain what the change is about and why it is desirable. -->

According to #13569 `parseString` is quite slow because it uses two loops scanning the input string byte-by-byte.

Replacing the loops by calls to `find` yields the main speedup of this PR.

## Context

<!-- Provide context. Reference open issues if available. -->

Even if the loops would be auto-vectorization friendly, auto-vectorization wouldn't be very effective because Nix is compiled for generic CPU targets and cannot use advanced instruction sets like AVX2 or AVX-512.
Standard libraries can implement variants for different instruction sets and can dispatch them dynamically. At least some generic, parallelized implementations of the most common string functions should be available.

With the PR the hello derivation is parsed ~7% faster, while the speedup of the firefox derivation benchmark is ~20% measured on Zen3.

The actual bottleneck in the firefox derivation benchmark isn't `parseString` anymore, but rather the [JSON parsing](https://github.com/NixOS/nix/blob/master/src/libstore/derivations.cc#L455) to `structuredAttrs`. Without actually parsing the JSON the throughput increases from ~117 MB/s to ~380 MB/s.
I'm not sure whether this code path is still relevant.

<!-- Non-trivial change: Briefly outline the implementation strategy. -->

<!-- Invasive change: Discuss alternative designs or approaches you considered. -->

<!-- Large change: Provide instructions to reviewers how to read the diff. -->

---

Add :+1: to [pull requests you find important](https://github.com/NixOS/nix/pulls?q=is%3Aopen+sort%3Areactions-%2B1-desc).

The Nix maintainer team uses a [GitHub project board](https://github.com/orgs/NixOS/projects/19) to [schedule and track reviews](https://github.com/NixOS/nix/tree/master/maintainers#project-board-protocol).