@xokdvium 2025-08-05
> Without actually parsing the JSON the throughput increases from ~117 MB/s to ~380 MB/s.

@Ericson2314, maybe doing the lazy parsing thing would be worthwhile actually?