@Mic92 2025-08-06
> > Without actually parsing the JSON the throughput increases from ~117 MB/s to ~380 MB/s.
> 
> @Ericson2314, maybe doing the lazy parsing thing would be worthwhile actually?

I did this [here](https://github.com/NixOS/nix/compare/master...Mic92:nix-1:lazy-json-drv?expand=1) and here is the performance diff:

```
# After
BM_ParseRealDerivationFile/hello        13473 ns        13260 ns        52542 bytes_per_second=126.873Mi/s
BM_ParseRealDerivationFile/firefox      46698 ns        46619 ns        15576 bytes_per_second=327.842Mi/s

# Before
BM_ParseRealDerivationFile/hello        14720 ns        14505 ns        47658 bytes_per_second=115.98Mi/s
BM_ParseRealDerivationFile/firefox     266988 ns       266251 ns         2608 bytes_per_second=57.4028Mi/s
```

Just unsure how relevant this is. This would need some benchmarks where we actually parse derivations.