@Ericson2314 2025-08-06
Yes, my take-away from the last time is while the JSON parsing is slow, derivation A-Term parsing performance also just doesn't matter than much, because GC doesn't (and shouldn't in the CA unstable feature case) need to parse drv files.

I think that means we can still do this PR, but it should be as much about code clarity as perform.