@NaN-git 2025-08-06
@Ericson2314 Are sure that the this code path is used only during GC? `Store::readDerivation` uses this code path and this function seems to have many callers. Or isn't the JSON parsed then?

I think that @Mic92's deferred computation of `structuredAttrs` would speed up many paths.